refactor: move channels_v2/capabilities.py to channels/capabilities.py
