+++ apps/channels/channels_v2/api_channel.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
-from apps.channels.channels_v2.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
 from apps.channels.channels_v2.sender import ChannelSender
