+++ apps/channels/channels_v2/channel_base.py
@@ -4,7 +4,7 @@ import logging
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, ClassVar
 
-from apps.channels.channels_v2.capabilities import ChannelCapabilities
+from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
 from apps.channels.channels_v2.stages.core import (
     AttachmentHydrationStage,
