+++ apps/channels/channels_v2/email_channel.py
@@ -11,8 +11,8 @@ from typing import TYPE_CHECKING
 from django.core.mail import EmailMessage as DjangoEmailMessage
 from django.db import IntegrityError
 
+from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.callbacks import ChannelCallbacks
-from apps.channels.channels_v2.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.const import MESSAGE_TYPES
